@@ -1,7 +1,9 @@
+# ruff: noqa
+# mypy: ignore-errors
+
 # Configuration file for the Sphinx documentation builder.
 #
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
+# For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
 # -- Path setup --------------------------------------------------------------
@@ -10,31 +12,35 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-import os
 import sys
+from pathlib import Path
 
-sys.path.insert(0, os.path.abspath("../../src"))
+sys.path.insert(0, str(Path("../../src").absolute()))
 
 
 # -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = "ospx"
 copyright = "2024, DNV SE. All rights reserved."
 author = "Frank Lumpitzsch, Claas Rostock, Seung Hyeon Yoo"
 
 # The full version, including alpha/beta/rc tags
-release = "0.2.14"
+release = "0.3.0b5"
 
 # -- General configuration ---------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
 extensions = [
     "myst_parser",
     "sphinx.ext.autodoc",
     "sphinx.ext.napoleon",
     "sphinx_argparse_cli",
+    "sphinx.ext.mathjax",
+    "matplotlib.sphinxext.plot_directive",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.todo",
+    "sphinxcontrib.mermaid",
 ]
 
 # The file extensions of source files.
@@ -51,24 +57,22 @@ templates_path = ["_templates"]
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = []
 
+
 # -- Options for HTML output -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-#
+html_title = f"ospx {release}"
 html_theme = "furo"
-html_logo = "_static/OSPx.svg"
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-
+html_logo = "_static/OSPx.svg"
 autodoc_default_options = {
-    "member-order": "bysource",
+    "member-order": "groupwise",
     "undoc-members": True,
     "exclude-members": "__weakref__",
 }
 autodoc_preserve_defaults = True
 
 myst_heading_anchors = 3
+
+# add markdown mermaid support
+myst_fence_as_directive = ["mermaid"]
