@@ -2,9 +2,9 @@ import logging
 import re
 from pathlib import Path
 from shutil import copy2
-from typing import Any, Dict, List, Union
+from typing import Any
 
-from dictIO import CppDict, DictWriter, XmlFormatter
+from dictIO import DictWriter, SDict, XmlFormatter
 from dictIO.utils.counter import BorgCounter
 from dictIO.utils.path import relative_path
 
@@ -21,10 +21,10 @@ class OspSimulationCase:
 
     def __init__(
         self,
-        case_dict: CppDict,
-    ):
+        case_dict: SDict[str, Any],
+    ) -> None:
         self.counter = BorgCounter()
-        self.case_dict: CppDict = case_dict
+        self.case_dict: SDict[str, Any] = case_dict
         self.case_folder: Path = case_dict.source_file.resolve().parent if case_dict.source_file else Path.cwd()
         self.system_structure: System
 
@@ -39,7 +39,7 @@ class OspSimulationCase:
         self.lib_source: Path
         self._resolve_lib_source_folder()
 
-    def setup(self):
+    def setup(self) -> None:
         """Set up the OSP simulation case folder.
 
         Raises
@@ -70,10 +70,14 @@ class OspSimulationCase:
         # Make sure all components have a step size defined
         self._check_components_step_size()
 
-    def _write_osp_model_description_xmls(self):
-        """Write the <component.name>_OspModelDescription.xml files for all components defined in the system structure."""
+    def _write_osp_model_description_xmls(self) -> None:
+        """Write the <component.name>_OspModelDescription.xml files for all components.
+
+        Writes the <component.name>_OspModelDescription.xml files for all components defined in the system structure.
+        """
         logger.info(
-            f"Write OspModelDescription.xml files for OSP simulation case '{self.name}' in case folder: {self.case_folder}"
+            f"Write OspModelDescription.xml files for OSP simulation case '{self.name}' "
+            f"in case folder: {self.case_folder}"
         )
         if not self.system_structure or not self.system_structure.components:
             return
@@ -81,7 +85,7 @@ class OspSimulationCase:
             component.write_osp_model_description_xml()
         return
 
-    def write_osp_system_structure_xml(self):
+    def write_osp_system_structure_xml(self) -> None:
         """Write the OspSystemStructure.xml file."""
         # sourcery skip: class-extract-method, merge-dict-assign
 
@@ -89,10 +93,11 @@ class OspSimulationCase:
         self._clean(osp_system_structure_file)
 
         logger.info(
-            f"Write OspSystemStructure.xml file for OSP simulation case '{self.name}' in case folder: {self.case_folder}"
+            f"Write OspSystemStructure.xml file for OSP simulation case '{self.name}' "
+            f"in case folder: {self.case_folder}"
         )
 
-        osp_system_structure: Dict[str, Any] = {}
+        osp_system_structure: dict[str, Any] = {}
         osp_system_structure["_xmlOpts"] = {
             "_nameSpaces": {"osp": "https://opensimulationplatform.com/xsd/OspModelDescription-1.0.0.xsd"},
             "_rootTag": "OspSystemStructure",
@@ -108,10 +113,10 @@ class OspSimulationCase:
                 osp_system_structure["Algorithm"] = self.simulation.algorithm
 
         # Simulators (=Components)
-        simulators: Dict[str, Any] = {}
-        for index, (_, component) in enumerate(self.system_structure.components.items()):
-            simulator_key = f"{index:06d}_Simulator"
-            simulator_properties: Dict[str, Dict[str, Union[str, float, Dict[str, Any], Path]]] = {
+        simulators: dict[str, Any] = {}
+        for i, (_, component) in enumerate(self.system_structure.components.items()):
+            simulator_key = f"{i:06d}_Simulator"
+            simulator_properties: dict[str, dict[str, str | float | dict[str, Any] | Path]] = {
                 "_attributes": {
                     "name": component.name,
                     "source": relative_path(self.case_folder, component.fmu.file),
@@ -130,16 +135,18 @@ class OspSimulationCase:
 
             if component.variables_with_start_values:
                 simulator_properties["InitialValues"] = {}
-                for index, (_, variable) in enumerate(component.variables_with_start_values.items()):
+                for ii, (_, variable) in enumerate(component.variables_with_start_values.items()):
                     if variable.start is not None and variable.data_type is None:
                         logger.error(
-                            f"component {component.name}: An initial value is defined for variable {variable.name}, but its data type is not defined.\n"
-                            f"The initial value for variable {variable.name} will not be written into OspSystemStructure.xml.\n"
+                            f"component {component.name}: An initial value is defined for variable {variable.name}, "
+                            "but its data type is not defined.\n"
+                            f"The initial value for variable {variable.name} will not be written "
+                            "into OspSystemStructure.xml.\n"
                             "OspSystemStructure.xml will be potentially wrong or incomplete."
                         )
                     else:
-                        initial_value_key = f"{index:06d}_InitialValue"
-                        initial_value_properties: Dict[str, Any] = {}
+                        initial_value_key = f"{ii:06d}_InitialValue"
+                        initial_value_properties: dict[str, Any] = {}
                         initial_value_properties["_attributes"] = {"variable": variable.name}
                         if variable.data_type:
                             initial_value_properties[variable.data_type] = {"_attributes": {"value": variable.start}}
@@ -150,7 +157,7 @@ class OspSimulationCase:
         osp_system_structure["Simulators"] = simulators
 
         # Connections
-        connections: Dict[str, Dict[str, Any]] = {}
+        connections: dict[str, dict[str, Any]] = {}
         for connection in self.system_structure.connections.values():
             if not connection.is_valid:
                 continue
@@ -202,9 +209,8 @@ class OspSimulationCase:
 
         return
 
-    def write_system_structure_ssd(self):
+    def write_system_structure_ssd(self) -> None:
         """Write the SystemStructure.ssd file."""
-
         system_structure_ssd_file = self.case_folder / "SystemStructure.ssd"
         self._clean(system_structure_ssd_file)
 
@@ -213,7 +219,7 @@ class OspSimulationCase:
             f"Write SystemStructure.ssd file for OSP simulation case '{self.name}' in case folder: {self.case_folder}"
         )
 
-        system_structure_ssd: Dict[str, Any] = {}
+        system_structure_ssd: dict[str, Any] = {}
         system_structure_ssd["_xmlOpts"] = {
             "_nameSpaces": {
                 "ssd": "file:///C:/Software/OSP/xsd/SystemStructureDescription",
@@ -249,9 +255,9 @@ class OspSimulationCase:
         system_structure_ssd["DefaultExperiment"] = default_experiment
 
         # Components
-        components: Dict[str, Any] = {}
+        components: dict[str, Any] = {}
         for component_name, component in self.system_structure.components.items():
-            connectors: Dict[str, Dict[str, Any]] = {}
+            connectors: dict[str, dict[str, Any]] = {}
             for connector in component.connectors.values():
                 if connector.variable and connector.type:
                     connector_key = f"{self.counter():06d}_Connector"
@@ -274,7 +280,7 @@ class OspSimulationCase:
         system_structure_ssd["System"]["Elements"] = components
 
         # Connections
-        connections: Dict[str, Any] = {}
+        connections: dict[str, Any] = {}
         for connection in self.system_structure.connections.values():
             if connection.source_endpoint and connection.target_endpoint:
                 connection_key = f"{self.counter():06d}_Connection"
@@ -294,18 +300,17 @@ class OspSimulationCase:
 
         return
 
-    def write_statistics_dict(self):
+    def write_statistics_dict(self) -> None:
         """Write selected properties of the system structure into a statistics dict.
 
         I.e. for documentation or further statistical analysis.
         """
         statistics_dict_file = self.case_folder / "statisticsDict"
-        # self._clean(statistics_dict_file)
 
         # sourcery skip: merge-dict-assign, simplify-dictionary-update
         logger.info(f"Write statistics dict for OSP simulation case '{self.name}' in case folder: {self.case_folder}")
 
-        statistics_dict = {}
+        statistics_dict: dict[str, dict[str, Any]] = {}
 
         statistics_dict["simulation"] = {"name": self.simulation.name}
 
@@ -324,10 +329,10 @@ class OspSimulationCase:
             "names": list(self.system_structure.connectors.keys()),
         }
 
-        unit_list: List[str] = []
-        display_unit_list: List[str] = []
-        factors_list: List[float] = []
-        offsets_list: List[float] = []
+        unit_list: list[str] = []
+        display_unit_list: list[str] = []
+        factors_list: list[float] = []
+        offsets_list: list[float] = []
         for unit in self.system_structure.units.values():
             unit_list.append(unit.name)
             display_unit_list.append(unit.display_unit.name)
@@ -349,7 +354,7 @@ class OspSimulationCase:
 
         DictWriter.write(statistics_dict, statistics_dict_file, mode="w")
 
-    def write_watch_dict(self):
+    def write_watch_dict(self) -> None:
         """Write a case-specific watch dict file.
 
         The watch dict file can be used with watchCosim for
@@ -358,11 +363,10 @@ class OspSimulationCase:
             - extracting the results
         """
         watch_dict_file = self.case_folder / "watchDict"
-        # self._clean(watch_dict_file)
 
         logger.info(f"Write watch dict for OSP simulation case '{self.name}' in case folder: {self.case_folder}")
 
-        watch_dict: Dict[str, Any] = {
+        watch_dict: dict[str, Any] = {
             "datasources": {},
             "delimiter": ",",  # 'objects': {},
             "simulation": {"name": self.simulation.name},
@@ -383,7 +387,7 @@ class OspSimulationCase:
 
         return
 
-    def _read_simulation(self):
+    def _read_simulation(self) -> None:
         """Read general simulation properties from case dict."""
         logger.info("reading simulation properties")  # 0
 
@@ -405,7 +409,7 @@ class OspSimulationCase:
             simulation.algorithm = simulation_properties["algorithm"]
         self.simulation = simulation
 
-    def _resolve_lib_source_folder(self):
+    def _resolve_lib_source_folder(self) -> None:
         """Resolve the library source folder."""
         self.lib_source = Path.cwd()  # initialize conservatively (with fallback path)
         if "_environment" in self.case_dict:
@@ -413,23 +417,22 @@ class OspSimulationCase:
                 self.lib_source = Path(self.case_dict["_environment"]["libSource"])
             else:
                 logger.warning(
-                    f"no 'libSource' element found in {self.case_dict.name}['_environment']. Path to libSource will be set to current working directory."
+                    f"no 'libSource' element found in {self.case_dict.name}['_environment']. "
+                    "Path to libSource will be set to current working directory."
                 )
         else:
             logger.warning(
-                f"no '_environment' section found in {self.case_dict.name}. Path to libSource hence is unknown and will be set to current working directory."
+                f"no '_environment' section found in {self.case_dict.name}. "
+                "Path to libSource hence is unknown and will be set to current working directory."
             )
         self.lib_source = self.lib_source.resolve().absolute()
 
     def _resolve_fmu_file(self, fmu_name: str) -> Path:
         fmu_file: Path = Path(fmu_name)
-        if fmu_file.is_absolute():
-            fmu_file = fmu_file.resolve()
-        else:
-            fmu_file = (self.lib_source / fmu_file).resolve()
+        fmu_file = fmu_file.resolve() if fmu_file.is_absolute() else (self.lib_source / fmu_file).resolve()
         return fmu_file
 
-    def _check_all_fmus_exist(self):
+    def _check_all_fmus_exist(self) -> None:
         """Check whether all referenced FMUs actually exist."""
         logger.debug("Check whether all referenced FMUs exist.")
         components = self.case_dict["systemStructure"]["components"]
@@ -445,7 +448,7 @@ class OspSimulationCase:
                 logger.exception(msg)
                 raise FileNotFoundError(fmu_file)
 
-    def _resolve_all_fmus(self):
+    def _resolve_all_fmus(self) -> None:
         """Resolve all referenced FMUs and ensures they are accessible from the case folder via a relative path.
 
         This is necessary because OspSystemStructure.xml allows only relative paths
@@ -453,10 +456,9 @@ class OspSimulationCase:
         If an FMU is not accessible via a relative path, the FMU will be copied into the case folder.
         Note: If multiple components reference the same FMU, these get copied only once.
         """
-
         logger.debug("Ensure all referenced FMUs are accessible from the case folder via a relative path.")
         components = self.case_dict["systemStructure"]["components"]
-        for _, component_properties in components.items():
+        for component_properties in components.values():
             fmu_file = self._resolve_fmu_file(component_properties["fmu"])
             try:
                 _ = relative_path(self.case_folder, fmu_file)
@@ -492,7 +494,7 @@ class OspSimulationCase:
                 copy2(osp_model_description_file, self.case_folder)
         return fmu_file_in_case_folder
 
-    def _check_components_step_size(self):
+    def _check_components_step_size(self) -> None:
         """Ensure that all components have a step size defined.
 
         If a components step size is undefined, it will be set to the base step size.
@@ -506,7 +508,7 @@ class OspSimulationCase:
                 component.step_size = self.simulation.base_step_size
         return
 
-    def _set_components_step_size(self, step_size: float):
+    def _set_components_step_size(self, step_size: float) -> None:
         """Overwrite the step size of all components with the passed in value."""
         if not self.system_structure or not self.system_structure.components:
             return
@@ -514,10 +516,11 @@ class OspSimulationCase:
             component.step_size = step_size
         return
 
-    def _inspect(self):
-        """Inspects all components and all FMUs for the public variable names and units they declare, as documented in their modelDescription.xml's.
+    def _inspect(self) -> None:
+        """Inspect all components and all FMUs for the public variable names and units they declare.
 
-        Results get logged to the console.
+        Inspects all components and all FMUs for the public variable names and units they declare
+        in their modelDescription.xml's. Results get logged to the console.
         """
         logger.info(f"Inspect OSP simulation case '{self.name}' in case folder: {self.case_folder}")
 
@@ -537,13 +540,13 @@ class OspSimulationCase:
                 f"\t{delim}{k}{delim}{v}" for k, v in fmu.model_description["_xmlOpts"]["_rootAttributes"].items()
             )
             log_string += fmu_attributes
-            if default_experiment_key := find_key(fmu.model_description, "DefaultExperiment$"):
-                if "_attributes" in fmu.model_description[default_experiment_key]:
-                    fmu_default_experiment = "\n".join(
-                        f"\t{delim}{k}{delim}{v}"
-                        for k, v in fmu.model_description[default_experiment_key]["_attributes"].items()
-                    )
-                    log_string += f"\n{fmu_default_experiment}"
+            default_experiment_key = find_key(fmu.model_description, "DefaultExperiment$")
+            if default_experiment_key and "_attributes" in fmu.model_description[default_experiment_key]:
+                fmu_default_experiment = "\n".join(
+                    f"\t{delim}{k}{delim}{v}"
+                    for k, v in fmu.model_description[default_experiment_key]["_attributes"].items()
+                )
+                log_string += f"\n{fmu_default_experiment}"
         logger.info(log_string + "\n")
 
         log_string = (
@@ -588,16 +591,15 @@ class OspSimulationCase:
 
         logger.info("Inspect mode: Finished.")
 
-    def _write_plot_config_json(self):
+    def _write_plot_config_json(self) -> None:
         """Write the PlotConfig.json file, containing postprocessing information."""
-
         plot_config_file = self.case_folder / "PlotConfig.json"
         self._clean(plot_config_file)
 
-        if "plots" in self.case_dict["postProcessing"].keys():
-            temp_dict: Dict[str, List[Dict[str, Any]]] = {"plots": []}
+        if "plots" in self.case_dict["postProcessing"]:
+            temp_dict: dict[str, list[dict[str, Any]]] = {"plots": []}
             for plot in self.case_dict["postproc"]["plots"].values():
-                variables: List[Dict[str, Any]] = []
+                variables: list[dict[str, Any]] = []
                 for (
                     component_name,
                     component,
@@ -623,20 +625,21 @@ class OspSimulationCase:
 
         return
 
-    def _correct_wrong_xml_namespace(self, file_name: str, pattern: str, replace: str):
-        """Substitutes namespace
-        (may be obsolete in future).
+    def _correct_wrong_xml_namespace(self, file_name: str, pattern: str, replace: str) -> None:
+        """Substitute namespace.
+
+        (may be obsolete in future)
         """
         buffer = ""
-        with open(file_name, "r") as f:
+        with Path(file_name).open() as f:
             buffer = re.sub(pattern, replace, f.read())
 
-        with open(file_name, "w") as f:
+        with Path(file_name).open(mode="w") as f:
             _ = f.write(buffer)
 
         return
 
-    def _clean(self, file_to_remove: Union[str, Path]):
+    def _clean(self, file_to_remove: str | Path) -> None:
         """Clean up single file."""
         if isinstance(file_to_remove, str):
             file_to_remove = self.case_folder / file_to_remove
