@@ -1,29 +1,33 @@
 import logging
 import os
-from glob import glob
 from pathlib import Path
 from shutil import rmtree
 
 import pytest
-from pytest import LogCaptureFixture
 
 from ospx.utils.zip import add_file_content_to_zip
 
 
 @pytest.fixture(scope="package", autouse=True)
-def chdir():
+def chdir() -> None:
+    """
+    Fixture that changes the current working directory to the 'test_working_directory' folder.
+    This fixture is automatically used for the entire package.
+    """
     os.chdir(Path(__file__).parent.absolute() / "test_dicts")
 
 
 @pytest.fixture(scope="package", autouse=True)
-def test_dir():
+def test_dir() -> Path:
+    """
+    Fixture that returns the absolute path of the directory containing the current file.
+    This fixture is automatically used for the entire package.
+    """
     return Path(__file__).parent.absolute()
 
 
-output_dirs = [
-    "xyz",
-]
-output_files = [
+output_dirs: list[str] = []
+output_files: list[str] = [
     "parsed*",
     "*.xml",
     "*.fmu",
@@ -36,7 +40,11 @@ output_files = [
 
 
 @pytest.fixture(autouse=True)
-def default_setup_and_teardown(caplog: LogCaptureFixture):
+def default_setup_and_teardown():
+    """
+    Fixture that performs setup and teardown actions before and after each test function.
+    It removes the output directories and files specified in 'output_dirs' and 'output_files' lists.
+    """
     _remove_output_dirs_and_files()
     _create_test_fmu()
     yield
@@ -44,20 +52,23 @@ def default_setup_and_teardown(caplog: LogCaptureFixture):
     _remove_output_dirs_and_files()
 
 
-def _remove_output_dirs_and_files():
+def _remove_output_dirs_and_files() -> None:
+    """
+    Helper function that removes the output directories and files specified in 'output_dirs' and 'output_files' lists.
+    """
     for folder in output_dirs:
         rmtree(folder, ignore_errors=True)
     for pattern in output_files:
-        for file in glob(pattern):
-            file = Path(file)
-            if not file.name.startswith("test_"):
-                file.unlink(missing_ok=True)
+        for file in Path.cwd().glob(pattern):
+            _file = Path(file)
+            if not _file.name.startswith("test_"):
+                _file.unlink(missing_ok=True)
 
 
-def _create_test_fmu():
+def _create_test_fmu() -> None:
     model_description_file: Path = Path("test_fmu_modelDescription.xml")
     model_description: str = ""
-    with open(model_description_file, "r") as f:
+    with Path.open(model_description_file) as f:
         model_description = f.read()
     fmu_file: Path = Path("test_fmu.fmu")
     fmu_file.unlink(missing_ok=True)
@@ -68,16 +79,21 @@ def _create_test_fmu():
     )
 
 
-def _remove_test_fmu():
+def _remove_test_fmu() -> None:
     Path("test_fmu.fmu").unlink()
 
 
 @pytest.fixture(autouse=True)
-def setup_logging(caplog: LogCaptureFixture):
-    caplog.set_level("WARNING")
+def setup_logging(caplog: pytest.LogCaptureFixture) -> None:
+    """
+    Fixture that sets up logging for each test function.
+    It sets the log level to 'INFO' and clears the log capture.
+    """
+    caplog.set_level("INFO")
     caplog.clear()
 
 
 @pytest.fixture(autouse=True)
-def logger():
+def logger() -> logging.Logger:
+    """Fixture that returns the logger object."""
     return logging.getLogger()
